+++ src/components/home/home.vue
